Fix math in doc, reorganize files and add loopless FVA options
